refactor: use standard project layout
