+++ internal/bobs/bobify.go
@@ -1,4 +1,4 @@
-package internal
+package bobs
 
 import (
 	"bufio"
