+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/tlkamp/mockbob/cmd"
+import "github.com/tlkamp/mockbob/internal/cmd"
 
 func main() {
 	cmd.Execute()
